Move apartment table to it's own component from ProjectCard
